move to loaders.py
